requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
aiomysql>=0.2.0
redis[hiredis]>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
asyncssh>=2.14.2
cryptography>=41.0.7
websockets>=12.0
jinja2>=3.1.2
bcrypt>=4.0.0,<5.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
python-a2s>=1.3.0
aiohttp>=3.9.4
captcha>=0.5.0
pillow>=10.3.0
